render alumni routes with logged in state check
